clean up and align all the bindings with index
